lib/keystore: migrate keyring to keystore
